requirements.txt
apache-airflow[amazon]==2.5.0
boto3==1.26.60
